requirements.txt
numpy==2.2.4
Flask
google-generativeai
python-dotenv
Werkzeug